requirements.txt
streamlit>=1.37
pandas>=2.2
requests>=2.32
python-dotenv>=1.0
PyYAML>=6.0
altair>=5.3
